6 sample optimization + 7 sample prep
I will copy and paste over to 7 sample once tuned; it doesn't make sense to copy over now. 6 sample original still uses some of the slow paths and less lenient end conditions, but added new sub intaking cases. Poses shouldn't need retuning. Just run a couple of times to test and tune timings.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystem/VerticalSlides.java
@@ -137,7 +137,7 @@ public class VerticalSlides {
     public void retract()           { moveToPosition(retractedPos);}
 //    public void raiseToPickupClip() { moveToPosition(pickupClipPos);}
 
-    public double getCurrentPos() { return currentPos;}
+    public int getCurrentPos() { return currentPos;}
 
     private boolean isDifferent(double val1, double val2) {
         return Math.abs(val1 - val2) >= CACHING_THRESHOLD;
